stock_forecast_project/requirements.txt
google-cloud-bigquery
yfinance
python-decouple
pandas
numpy
fbprophet
plotly
